move hook to hooks folder
